search functionality

+++ app/Http/Controllers/API/UserController.php
@@ -26,7 +26,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(20);
+        // $this->authorize('isAdmin');
+        if( \Gate::allows('isAdmin') || \Gate::allows('isAuthor') ){
+            return User::latest()->paginate(20);
+
+        }
+
     }
 
     /**
@@ -125,10 +130,28 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('isAdmin');
+
         $user = User::findOrFail($id);
 
         // delete
         $user->delete();
         return ['message' => "User Deleted"];
     }
+    
+    public function search(){
+        if ($search = \Request::get('q')) {
+            $users = User::where(function($query) use ($search){
+                $query->where('name','LIKE',"%$search%")
+                        ->orWhere('email','LIKE',"%$search%");
+            })->paginate(20);
+        }else{
+            $users = User::latest()->paginate(5);
+        }
+        return $users;
+    }
+
+
+
+
 }
